server/requirements.txt
flask
pandas
python_dotenv
pymongo
pyjwt
gunicorn
flask_cors
bcrypt
nanoid
gunicorn
google.generativeai
b2sdk
PyPDF2
pika
redis
